This is a collection of hacked-up and pared-down tests
and debug logging that reproduces a corrupt state in
the Gem class.  See comment in spec/spec_helper.rb labeled
'Hack warning' for all the gory details.

In master, this is temporarily worked around by enabling the
Gem.clear_paths in config.before(:each); that is commented out in this
branch so that it can be repro'd immediately.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>